Move the snapshot handling into main.qml
Press 'S' to grab the screen including any popups etc. (useful for creating
images for documentation. This is available anywhere. The image will be saved
to {configdir}/snapshots/ (eg. ~/.mythtv/snapshots/snapshot0001.png)

When the InternalPlayer is active press 'F11' to grab a snapshot of the active
video player (useful to grab thumbnail images for videos). If the playing
video is a local file (ie the URL starts with 'file://') then the snapshot
will be saved to the same directory as the video and the filename will be
the same as the video file with '.png' appended otherwise it will be saved
to the snapshots directory like above.