Add new helper app called mythlauncher_qml
This new app is designed to be the app shown at startup and allows you to
choose which frontend you want to use. Currently they are the old frontend
the new QML frontend or kodi.

By default the launcher will auto start the new QML frontend at startup.

There are additional options to exit the app, shutdown or reboot the system
and change settings.