MediaPlayer: add support for displaying websites in the left browser
Webcam feeds can now have up to 10 website pages defined in trhe media feed
which can be displayed in the brower while watching a feed.

Use the menu to choose which feed to see or RED/GREEN to switch to the previous
or next feed while thge brower has focus.