ZoneMinder: add event status monitoring
This allows the video player to show what state a camera is in for example
it will show Alarm if a camera is monitoring for motion and an alarm has
been triggered.

The events list view will now default to showing only non archived events
which is what you want 99% of the time. Press BLUE (F4) to show archived
or all events.

There is also some preliminary support for triggering and cancelling events
manually.