Add option to change the fill mode of the players
This option changes the way the video is displayed in the players. You can
choose to stretch the video, shrink to fit maintaining the aspect ratio,
enlarge to fit cropping as required.

Press 'A' to toggle the available fill modes.

Currently the QtAV, VLC and MDK players support the new option.